# PERF: read_stata