```
from asv_bench.benchmarks.io.stata import *

self = StataMissing()
self.setup("td")

%timeit self.time_read_stata("td")
113 ms ± 1.44 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- master
94.4 ms ± 1.72 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- PR
```

cc @bashtage